CppLintParser: Allow double colon in message
Require the line number to be at least one digit by changing from "*" to
"+". This will prevent the parser from treating the C++ scope resolution
operator (::) as the separator between the filename, line number and
message.

The problem was that in the following warning:
"path/to/file:12: warning with std::string [type] [2]"
the filename regex (.*) would be so greedy that it included everything
up to "std" and set the line number to an empty string.