[JENKINS-56393][JENKINS-56394] RfLintParser improvements
Change Description:
The RfLintParser was not displaying the correct category.  It was using the
severity as the category, instead of the rule name. In addition, an absolute
path for the directory and was not display the package name.  Although robot
does not have the concept of package, the same filename can exist is multiple
directories.  As a such, it was difficult to know which specific file has the
violation.

The following changes are introduced:

1. When a violation/issue is reported, the category is set to one of:
   -Suite
   - Test Case
   - Keyword
   - Other
   -Custom
2. When a violation/issue is reported, the type is set as the rule name
3. Set the directory name so the same path as reported in the RfLint log file
   is displayed
4. Set the package name to the files parent folder

Testing Done:
1. Locally build the analysis-model project using `man clean install`
2. Replaced ${JENKINS_HOME}/plugins/analysis-model-api/WEB-INF/lib/analysis-model-x.y.x.jar
   with <repo>/target/analysis-model-<version>.jar
3. Trigger a job and ensure the categories are updated correctly