56737 - Allow timestamps in front of javac warnings
Javac warnings were not matched if a timestamp preceded the message.
Example:
08:52:34.395 [WARNING] /my/path/to/workspace/src/main/.....

Warnings with timestamps will now be matched. The preceeding word has to
to be a arbitrary number of characters that are no whitespace followed
by one whitespace. It is not checked if the preceding word is a
timestamp at all because of the varous timestamp formats that are allowed.