Resolve some path normalization issues
When the tests stub workspace project paths to return
`PATH_PREFIX_FOO + ModuleDetector.FOO`, the string is getting mangled to
remove the path nature of the string: "path/to/mavenpom.xml", but we need the
path to be "path/to/maven/pom.xml".

To fix that, and following the existing example of PREFIX ending in a path-
separator, this fixes the PATH_PREFIX_* constants to each end with a path-
separator, and removes the path separator from the appended strings.