Detect Gradle module names
This is a primitive detection that assumes a Gradle's module name is equal to
the directory name of the closest build.gradle file.

For example:

    workspace=/path/to/gradle-project/
    build script=/path/to/gradle-project/build.gradle
    report file=/path/to/gradle-project/build/reports/something.txt

would resolve this report to the parent directory name of the build.gradle
file (in this case, `gradle-project`).

This works well for the default gradle project structure, but will need some
tweaking in case of the root module being overridden by settings.gradle.