Write the latest version of the library as property into the POM
RevApi compares our code to the latest released version of this library.
To find this version, it depends on the artifacts deployed to Maven
repositories. Unfortunately, the analysis-model is released in two
different repositories, making it impossible to auto-detect.
Therefore, we need to write the latest version into the POM
(during the release process) and provide this version as property
for RevApi.