[JENKINS-70855] Skip duplicate warnings from imported header files
When a warning is identified in a C/C++ header file and this file is
included in several source code files, then the warning appears
for each distinct import. It makes more sense to skip those
duplicates, and count such warnings only once.

We use a pragmatic fix to provide a solution for JENKINS-70855: we
simply remove the offending "import" text to make the warnings equal
to each other.

See https://issues.jenkins.io/browse/JENKINS-70855